* opengeek/impr-cache: (46 commits)
  Remove resourceListing, documentListing, and documentMap from context cache
  Additional settings and lexicons for cache refactoring
  Preserve config_options in config file during setup
  Avoid PHP warning from fopen in xPDOFileCache->get()
  Bump version for 2.1.0-pl2
  Mute warnings/errors from fopen/fclose in writeFile()
  Bump version for 2.1.0 release
  Add proper support for cache file formats other than PHP includes
  Add proper cache file locking and format handling
  Add OPT_CACHE_FORMAT constant to xPDO
  Cleanup clearcache output from context_settings
  Allow partition-specific cache expiration settings
  Fix generateActionMap
  Store/re-use parent paths as they are calculated when generating context cache
  Refactor pluginCache generation
  Get rid of aliasListing
  Use modX::$_systemConfig for context settings generation
  Add missing partitions for system/context_settings
  Add CACHE_SERIALIZE format support
  Additional modCacheManager->refresh() refactoring
  ...

+++ core/xpdo/cache/xpdocachemanager.class.php
@@ -34,6 +34,7 @@
 class xPDOCacheManager {
     const CACHE_PHP = 0;
     const CACHE_JSON = 1;
+    const CACHE_SERIALIZE = 2;
     const CACHE_DIR = 'objects/';
     const LOG_DIR = 'logs/';
 
@@ -194,24 +195,30 @@ class xPDOCacheManager {
      * @param string $filename The absolute path to the location the file will
      * be written in.
      * @param string $content The content of the newly written file.
-     * @param string $mode The php file mode to write in. Defaults to 'wb'
+     * @param string $mode The php file mode to write in. Defaults to 'ab'
      * @param array $options An array of options for the function.
      * @return boolean Returns true if the file was successfully written.
      */
-    public function writeFile($filename, $content, $mode= 'wb', $options= array()) {
+    public function writeFile($filename, $content, $mode= 'ab', $options= array()) {
         $written= false;
-        if (!is_array($options)) $options = is_scalar($options) && !is_bool($options) ? array('new_folder_permissions' => $options) : array();
+        if (!is_array($options)) {
+            $options = is_scalar($options) && !is_bool($options) ? array('new_folder_permissions' => $options) : array();
+        }
         $dirname= dirname($filename);
         if (!file_exists($dirname)) {
-            if ($this->writeTree($dirname, $options)) {
-                $file= @ fopen($filename, $mode);
+            $this->writeTree($dirname, $options);
+        }
+        $file= @fopen($filename, $mode);
+        if ($file) {
+            if (flock($file, LOCK_EX | LOCK_NB)) {
+                fseek($file, 0);
+                ftruncate($file, 0);
+                $written= fwrite($file, $content);
+                flock($file, LOCK_UN);
             }
+            @fclose($file);
         }
-        if ($file= @ fopen($filename, $mode)) {
-            $written= @ fwrite($file, $content);
-            @ fclose($file);
-        }
-        return $written;
+        return ($written !== false);
     }
 
     /**
@@ -652,6 +659,30 @@ class xPDOCacheManager {
     }
 
     /**
+     * Refresh specific or all cache providers.
+     *
+     * The default behavior is to call clean() with the provided options
+     *
+     * @param array $providers An associative array with keys representing the cache provider key
+     * and the value an array of options.
+     * @param array &$results An associative array for collecting results for each provider.
+     * @return array An array of results for each provider that is refreshed.
+     */
+    public function refresh(array $providers = array(), array &$results = array()) {
+        if (empty($providers)) {
+            foreach ($this->caches as $cacheKey => $cache) {
+                $providers[$cacheKey] = array();
+            }
+        }
+        foreach ($providers as $key => $options) {
+            if (array_key_exists($key, $this->caches) && !array_key_exists($key, $results)) {
+                $results[$key] = $this->clean(array_merge($options, array(xPDO::OPT_CACHE_KEY => $key)));
+            }
+        }
+        return (array_search(false, $results, true) === false);
+    }
+
+    /**
      * Escapes all single quotes in a string
      *
      * @access public
@@ -740,7 +771,7 @@ abstract class xPDOCache {
     public function getCacheKey($key, $options = array()) {
         $prefix = $this->getOption('cache_prefix', $options);
         if (!empty($prefix)) $key = $prefix . $key;
-        return $key;
+        return $this->key . '/' . $key;
     }
 
     /**
@@ -854,11 +885,19 @@ class xPDOFileCache extends xPDOCache {
                 $expireContent= 'if(time() > ' . $expirationTS . '){return null;}';
             }
             $fileName= $this->getCacheKey($key, $options);
-            if (!empty($options['format']) && $options['format'] == xPDOCacheManager::CACHE_JSON) {
-                $content= !is_scalar($var) ? $this->xpdo->toJSON($var) : $var;
-            } else {
-            $content= '<?php ' . $expireContent . ' return ' . var_export($var, true) . ';';
-        }
+            $format = (integer) $this->getOption(xPDO::OPT_CACHE_FORMAT, $options, xPDOCacheManager::CACHE_PHP);
+            switch ($format) {
+                case xPDOCacheManager::CACHE_SERIALIZE:
+                    $content= serialize(array('expires' => $expirationTS, 'content' => $var));
+                    break;
+                case xPDOCacheManager::CACHE_JSON:
+                    $content= $this->xpdo->toJSON(array('expires' => $expirationTS, 'content' => $var));
+                    break;
+                case xPDOCacheManager::CACHE_PHP:
+                default:
+                    $content= '<?php ' . $expireContent . ' return ' . var_export($var, true) . ';';
+                    break;
+            }
             $set= $this->xpdo->cacheManager->writeFile($fileName, $content);
         }
         return $set;
@@ -878,9 +917,12 @@ class xPDOFileCache extends xPDOCache {
         $deleted= false;
         $cacheKey= $this->getCacheKey($key, array_merge($options, array('cache_ext' => '')));
         if (file_exists($cacheKey) && is_dir($cacheKey)) {
-            $deleted= $this->xpdo->cacheManager->deleteTree($cacheKey, false, true);
+            $results = $this->xpdo->cacheManager->deleteTree($cacheKey, array_merge(array('deleteTop' => false, 'skipDirs' => false, 'extensions' => array('.cache.php')), $options));
+            if ($results !== false) {
+                $deleted = true;
+            }
         } else {
-            $cacheKey.= $this->getOption('cache_ext', $options, '.cache.php');
+            $cacheKey= $this->getCacheKey($key, $options);
             if (file_exists($cacheKey)) {
                 $deleted= @ unlink($cacheKey);
             }
@@ -892,13 +934,44 @@ class xPDOFileCache extends xPDOCache {
         $value= null;
         $cacheKey= $this->getCacheKey($key, $options);
         if (file_exists($cacheKey)) {
-            if (!empty($options['format']) && $options['format'] == xPDOCacheManager::CACHE_JSON) {
-                $value= file_get_contents($cacheKey);
-            } else {
-                $value= @ include ($cacheKey);
-            }
-            if ($value === null && $this->getOption('removeIfEmpty', $options, true)) {
-                @ unlink($cacheKey);
+            if ($file = @fopen($cacheKey, 'rb')) {
+                $format = (integer) $this->getOption(xPDO::OPT_CACHE_FORMAT, $options, xPDOCacheManager::CACHE_PHP);
+                if (flock($file, LOCK_SH)) {
+                    switch ($format) {
+                        case xPDOCacheManager::CACHE_PHP:
+                            $value= @include $cacheKey;
+                            break;
+                        case xPDOCacheManager::CACHE_JSON:
+                            $payload = stream_get_contents($file);
+                            if ($payload !== false) {
+                                $payload = $this->xpdo->fromJSON($payload);
+                                if (is_array($payload) && isset($payload['expires']) && (empty($payload['expires']) || time() < $payload['expires'])) {
+                                    if (array_key_exists('content', $payload)) {
+                                        $value= $payload['content'];
+                                    }
+                                }
+                            }
+                            break;
+                        case xPDOCacheManager::CACHE_SERIALIZE:
+                            $payload = stream_get_contents($file);
+                            if ($payload !== false) {
+                                $payload = unserialize($payload);
+                                if (is_array($payload) && isset($payload['expires']) && (empty($payload['expires']) || time() < $payload['expires'])) {
+                                    if (array_key_exists('content', $payload)) {
+                                        $value= $payload['content'];
+                                    }
+                                }
+                            }
+                            break;
+                    }
+                    flock($file, LOCK_UN);
+                    if ($value === null && $this->getOption('removeIfEmpty', $options, true)) {
+                        fclose($file);
+                        @ unlink($cacheKey);
+                        return $value;
+                    }
+                }
+                @fclose($file);
             }
         }
         return $value;
@@ -906,6 +979,7 @@ class xPDOFileCache extends xPDOCache {
 
     public function flush($options= array()) {
         $cacheKey= $this->getCacheKey('', array_merge($options, array('cache_ext' => '')));
-        return $this->xpdo->cacheManager->deleteTree($cacheKey, false, true);
+        $results = $this->xpdo->cacheManager->deleteTree($cacheKey, array_merge(array('deleteTop' => false, 'skipDirs' => false, 'extensions' => array('.cache.php')), $options));
+        return ($results !== false);
     }
 }
